Split orchestrator phase helpers into modules
