+++ crates/xchecker-engine/src/orchestrator/mod.rs
@@ -7,6 +7,8 @@
 mod handle;
 mod llm;
 mod phase_exec;
+mod phase_factory;
+mod simulation;
 mod workflow;
 
 #[allow(unused_imports)]
